+++ src/coker/dynamics/analysis/__init__.py
@@ -1,9 +1,8 @@
-"""Generic local symbolic analysis for Coker dynamical systems."""
+"""Structural analysis of Coker dynamical systems."""
 
 from .controllability import analyse_controllability
-from .identifiability import analyse_identifiability
-
 from .dae import SymbolicDAESystem, lower_dae_system
+from .identifiability import analyse_identifiability
 from .model import AnalysisResult, AnalysisStatus
 from .symbolic import SymbolicSystem, UnsupportedSystemError, lower_system
 
@@ -15,6 +14,6 @@ __all__ = (
     "UnsupportedSystemError",
     "analyse_controllability",
     "analyse_identifiability",
-    "lower_system",
     "lower_dae_system",
+    "lower_system",
 )